it works... it can be used to play the game now. Timer automatically launched when the second view controller is viewed. Shift view controller automatically launches when timer fires. Shift view controller times out after 15 seconds or the button can be pressed to return to the second view controller, upon which the timer starts again.
tested for bugs by changing view controllers when the timer is supposed to launch. i was prompted that this is not a good idea but it worked fine.
removed players mid game and game runs as expected after adjusting the player number.

i would like to make the animation change each time so that each image view animation duration is random.
i tried to add sound for when timer fires. sound played but i could not get it to loop and then end when shift view is dismissed.
still want to have plist files for previous players in the game.
and to adapt the program to be able to work on iPad without it looking terrible... which it does.
background image my change but I'm not worried about that.
still want to as swipe or maybe just a start button to begin game after players are added, and then remove the tab bar. I'm not sure if that will break the program because the tab controller is managing the flow i think.

other than that its mostly cosmetics and twerk the time intervals when things happen. But it is playable, need to try it soon!!!